backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python-multipart>=0.0.6
pymongo>=4.3.3
motor>=3.1.1
python-dotenv>=1.0.0
httpx>=0.24.0
pandas>=2.0.0
openai>=1.14.0 